fix Header and Dir Structure in Event
