+++ srcs/Event/EventBase/EventFactory.cpp
@@ -1,0 +1,1 @@
+#include "../../../incs/Event/EventBase/EventFactory.hpp"
