+++ srcs/Event/EventBase/EventHandler.cpp
@@ -1,0 +1,4 @@
+#include "../../../incs/Event/EventBase/EventHandler.hpp"
+
+EventHandler::EventHandler(void) {}
+EventHandler::~EventHandler(void) {}
