+++ srcs/Event/ReadEvent/ReadEventHandler.cpp
@@ -1,0 +1,4 @@
+#include "../../../incs/Event/ReadEvent/ReadEventHandler.hpp"
+
+ReadEventHandler::ReadEventHandler() {}
+ReadEventHandler::~ReadEventHandler() {}
