--- srcs/Event/TimeEvent.cpp
@@ -1,7 +1,0 @@
-#include "../../incs/Event/TimeEvent.hpp"
-#include <new>
-
-// To do: implement TimeEvent
-
-//TimeEvent::TimeEvent(void): Event() {}
-//TimeEvent::~TimeEvent(void) {}
